editors - introduce a `AbstractTextResourceEditorInput` for some label sharing
