+++ src/vs/workbench/browser/parts/editor/textResourceEditor.ts
@@ -7,7 +7,7 @@ import { localize } from 'vs/nls';
 import { assertIsDefined, isFunction, withNullAsUndefined } from 'vs/base/common/types';
 import { ICodeEditor, getCodeEditor, IPasteEvent } from 'vs/editor/browser/editorBrowser';
 import { TextEditorOptions, EditorInput, EditorOptions, IEditorOpenContext } from 'vs/workbench/common/editor';
-import { ResourceEditorInput } from 'vs/workbench/common/editor/resourceEditorInput';
+import { TextResourceEditorInput } from 'vs/workbench/common/editor/textResourceEditorInput';
 import { BaseTextEditorModel } from 'vs/workbench/common/editor/textEditorModel';
 import { UntitledTextEditorInput } from 'vs/workbench/services/untitled/common/untitledTextEditorInput';
 import { BaseTextEditor } from 'vs/workbench/browser/parts/editor/textEditor';
@@ -98,7 +98,7 @@ export class AbstractTextResourceEditor extends BaseTextEditor {
 	}
 
 	private restoreTextResourceEditorViewState(editor: EditorInput, control: IEditor) {
-		if (editor instanceof UntitledTextEditorInput || editor instanceof ResourceEditorInput) {
+		if (editor instanceof UntitledTextEditorInput || editor instanceof TextResourceEditorInput) {
 			const viewState = this.loadTextEditorViewState(editor.resource);
 			if (viewState) {
 				control.restoreViewState(viewState);
@@ -144,7 +144,7 @@ export class AbstractTextResourceEditor extends BaseTextEditor {
 	}
 
 	private saveTextResourceEditorViewState(input: EditorInput | undefined): void {
-		if (!(input instanceof UntitledTextEditorInput) && !(input instanceof ResourceEditorInput)) {
+		if (!(input instanceof UntitledTextEditorInput) && !(input instanceof TextResourceEditorInput)) {
 			return; // only enabled for untitled and resource inputs
 		}
