+++ src/vs/workbench/common/editor/resourceEditorInput.ts
@@ -3,132 +3,194 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { localize } from 'vs/nls';
+import { EditorInput, Verbosity, IEditorInputWithPreferredResource } from 'vs/workbench/common/editor';
 import { URI } from 'vs/base/common/uri';
-import { IReference } from 'vs/base/common/lifecycle';
-import { ITextEditorModel, ITextModelService } from 'vs/editor/common/services/resolverService';
-import { ResourceEditorModel } from 'vs/workbench/common/editor/resourceEditorModel';
-import { IModeSupport, ITextFileService } from 'vs/workbench/services/textfile/common/textfiles';
-import { IEditorService } from 'vs/workbench/services/editor/common/editorService';
-import { IEditorGroupsService } from 'vs/workbench/services/editor/common/editorGroupsService';
-import { IFileService } from 'vs/platform/files/common/files';
+import { IFileService, FileSystemProviderCapabilities } from 'vs/platform/files/common/files';
 import { ILabelService } from 'vs/platform/label/common/label';
-import { IFilesConfigurationService } from 'vs/workbench/services/filesConfiguration/common/filesConfigurationService';
-import { AbstractTextResourceEditorInput } from 'vs/workbench/common/editor/textResourceEditorInput';
-import { isEqual } from 'vs/base/common/resources';
+import { dirname, isEqual } from 'vs/base/common/resources';
 
 /**
- * A read-only text editor input whos contents are made of the provided resource that points to an existing
- * code editor model.
+ * The base class for all editor inputs that open resources.
  */
-export class ResourceEditorInput extends AbstractTextResourceEditorInput implements IModeSupport {
+export abstract class AbstractResourceEditorInput extends EditorInput implements IEditorInputWithPreferredResource {
 
-	static readonly ID: string = 'workbench.editors.resourceEditorInput';
-
-	override get typeId(): string {
-		return ResourceEditorInput.ID;
-	}
-
-	private cachedModel: ResourceEditorModel | undefined = undefined;
-	private modelReference: Promise<IReference<ITextEditorModel>> | undefined = undefined;
+	private _preferredResource: URI;
+	get preferredResource(): URI { return this._preferredResource; }
 
 	constructor(
-		resource: URI,
-		private name: string | undefined,
-		private description: string | undefined,
-		private preferredMode: string | undefined,
-		@ITextModelService private readonly textModelResolverService: ITextModelService,
-		@ITextFileService textFileService: ITextFileService,
-		@IEditorService editorService: IEditorService,
-		@IEditorGroupsService editorGroupService: IEditorGroupsService,
-		@IFileService fileService: IFileService,
-		@ILabelService labelService: ILabelService,
-		@IFilesConfigurationService filesConfigurationService: IFilesConfigurationService
+		public readonly resource: URI,
+		preferredResource: URI | undefined,
+		@ILabelService protected readonly labelService: ILabelService,
+		@IFileService protected readonly fileService: IFileService
 	) {
-		super(resource, undefined, editorService, editorGroupService, textFileService, labelService, fileService, filesConfigurationService);
+		super();
+
+		this._preferredResource = preferredResource || resource;
+
+		this.registerListeners();
 	}
 
-	override getName(): string {
-		return this.name || super.getName();
+	private registerListeners(): void {
+
+		// Clear label memoizer on certain events that have impact
+		this._register(this.labelService.onDidChangeFormatters(e => this.onLabelEvent(e.scheme)));
+		this._register(this.fileService.onDidChangeFileSystemProviderRegistrations(e => this.onLabelEvent(e.scheme)));
+		this._register(this.fileService.onDidChangeFileSystemProviderCapabilities(e => this.onLabelEvent(e.scheme)));
 	}
 
-	setName(name: string): void {
-		if (this.name !== name) {
-			this.name = name;
-
-			this._onDidChangeLabel.fire();
+	private onLabelEvent(scheme: string): void {
+		if (scheme === this._preferredResource.scheme) {
+			this.updateLabel();
 		}
 	}
 
-	override getDescription(): string | undefined {
-		return this.description;
+	private updateLabel(): void {
+
+		// Clear any cached labels from before
+		this._name = undefined;
+		this._shortDescription = undefined;
+		this._mediumDescription = undefined;
+		this._longDescription = undefined;
+		this._shortTitle = undefined;
+		this._mediumTitle = undefined;
+		this._longTitle = undefined;
+
+		// Trigger recompute of label
+		this._onDidChangeLabel.fire();
 	}
 
-	setDescription(description: string): void {
-		if (this.description !== description) {
-			this.description = description;
+	setPreferredResource(preferredResource: URI): void {
+		if (!isEqual(preferredResource, this._preferredResource)) {
+			this._preferredResource = preferredResource;
 
-			this._onDidChangeLabel.fire();
+			this.updateLabel();
 		}
 	}
 
-	setMode(mode: string): void {
-		this.setPreferredMode(mode);
+	private _name: string | undefined = undefined;
+	override getName(skipDecorate?: boolean): string {
+		if (typeof this._name !== 'string') {
+			this._name = this.labelService.getUriBasenameLabel(this._preferredResource);
+		}
 
-		if (this.cachedModel) {
-			this.cachedModel.setMode(mode);
+		return skipDecorate ? this._name : this.decorateLabel(this._name);
+	}
+
+	override getDescription(verbosity = Verbosity.MEDIUM): string | undefined {
+		switch (verbosity) {
+			case Verbosity.SHORT:
+				return this.shortDescription;
+			case Verbosity.LONG:
+				return this.longDescription;
+			case Verbosity.MEDIUM:
+			default:
+				return this.mediumDescription;
 		}
 	}
 
-	setPreferredMode(mode: string): void {
-		this.preferredMode = mode;
+	private _shortDescription: string | undefined = undefined;
+	private get shortDescription(): string {
+		if (typeof this._shortDescription !== 'string') {
+			this._shortDescription = this.labelService.getUriBasenameLabel(dirname(this._preferredResource));
+		}
+
+		return this._shortDescription;
 	}
 
-	override async resolve(): Promise<ITextEditorModel> {
-		if (!this.modelReference) {
-			this.modelReference = this.textModelResolverService.createModelReference(this.resource);
+	private _mediumDescription: string | undefined = undefined;
+	private get mediumDescription(): string {
+		if (typeof this._mediumDescription !== 'string') {
+			this._mediumDescription = this.labelService.getUriLabel(dirname(this._preferredResource), { relative: true });
 		}
 
-		const ref = await this.modelReference;
-
-		// Ensure the resolved model is of expected type
-		const model = ref.object;
-		if (!(model instanceof ResourceEditorModel)) {
-			ref.dispose();
-			this.modelReference = undefined;
-
-			throw new Error(`Unexpected model for ResourceEditorInput: ${this.resource}`);
-		}
-
-		this.cachedModel = model;
-
-		// Set mode if we have a preferred mode configured
-		if (this.preferredMode) {
-			model.setMode(this.preferredMode);
-		}
-
-		return model;
+		return this._mediumDescription;
 	}
 
-	override matches(otherInput: unknown): boolean {
-		if (otherInput === this) {
-			return true;
+	private _longDescription: string | undefined = undefined;
+	private get longDescription(): string {
+		if (typeof this._longDescription !== 'string') {
+			this._longDescription = this.labelService.getUriLabel(dirname(this._preferredResource));
 		}
 
-		if (otherInput instanceof ResourceEditorInput) {
-			return isEqual(otherInput.resource, this.resource);
+		return this._longDescription;
+	}
+
+	private _shortTitle: string | undefined = undefined;
+	private get shortTitle(): string {
+		if (typeof this._shortTitle !== 'string') {
+			this._shortTitle = this.getName(true /* skip decorations */);
 		}
 
+		return this._shortTitle;
+	}
+
+	private _mediumTitle: string | undefined = undefined;
+	private get mediumTitle(): string {
+		if (typeof this._mediumTitle !== 'string') {
+			this._mediumTitle = this.labelService.getUriLabel(this._preferredResource, { relative: true });
+		}
+
+		return this._mediumTitle;
+	}
+
+	private _longTitle: string | undefined = undefined;
+	private get longTitle(): string {
+		if (typeof this._longTitle !== 'string') {
+			this._longTitle = this.labelService.getUriLabel(this._preferredResource);
+		}
+
+		return this._longTitle;
+	}
+
+	override getTitle(verbosity?: Verbosity): string {
+		switch (verbosity) {
+			case Verbosity.SHORT:
+				return this.decorateLabel(this.shortTitle);
+			case Verbosity.LONG:
+				return this.decorateLabel(this.longTitle);
+			default:
+			case Verbosity.MEDIUM:
+				return this.decorateLabel(this.mediumTitle);
+		}
+	}
+
+	private decorateLabel(label: string): string {
+		const readonly = this.isReadonly();
+		const orphaned = this.isOrphaned();
+
+		return decorateFileEditorLabel(label, { orphaned, readonly });
+	}
+
+	override isUntitled(): boolean {
+		return !this.fileService.canHandleResource(this.resource);
+	}
+
+	override isReadonly(): boolean {
+		if (this.isUntitled()) {
+			return false; // untitled is never readonly
+		}
+
+		return this.fileService.hasCapability(this.resource, FileSystemProviderCapabilities.Readonly);
+	}
+
+	isOrphaned(): boolean {
 		return false;
 	}
+}
 
-	override dispose(): void {
-		if (this.modelReference) {
-			this.modelReference.then(ref => ref.dispose());
-			this.modelReference = undefined;
-		}
-
-		this.cachedModel = undefined;
-
-		super.dispose();
+export function decorateFileEditorLabel(label: string, state: { orphaned: boolean, readonly: boolean }): string {
+	if (state.orphaned && state.readonly) {
+		return localize('orphanedReadonlyFile', "{0} (deleted, read-only)", label);
 	}
+
+	if (state.orphaned) {
+		return localize('orphanedFile', "{0} (deleted)", label);
+	}
+
+	if (state.readonly) {
+		return localize('readonlyFile', "{0} (read-only)", label);
+	}
+
+	return label;
 }
