+++ src/vs/workbench/common/editor/textResourceEditorModel.ts
@@ -9,9 +9,10 @@ import { IModeService } from 'vs/editor/common/services/modeService';
 import { IModelService } from 'vs/editor/common/services/modelService';
 
 /**
- * An editor model for in-memory, readonly content that is backed by an existing editor model.
+ * An editor model for in-memory, readonly text content that
+ * is backed by an existing editor model.
  */
-export class ResourceEditorModel extends BaseTextEditorModel {
+export class TextResourceEditorModel extends BaseTextEditorModel {
 
 	constructor(
 		resource: URI,
