+++ src/vs/workbench/contrib/notebook/common/notebookEditorInput.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as glob from 'vs/base/common/glob';
-import { EditorInput, IEditorInput, GroupIdentifier, ISaveOptions, IMoveResult, IRevertOptions } from 'vs/workbench/common/editor';
+import { IEditorInput, GroupIdentifier, ISaveOptions, IMoveResult, IRevertOptions } from 'vs/workbench/common/editor';
 import { INotebookService } from 'vs/workbench/contrib/notebook/common/notebookService';
 import { URI } from 'vs/base/common/uri';
 import { isEqual, joinPath } from 'vs/base/common/resources';
@@ -16,12 +16,14 @@ import { IResolvedNotebookEditorModel } from 'vs/workbench/contrib/notebook/comm
 import { ILabelService } from 'vs/platform/label/common/label';
 import { Schemas } from 'vs/base/common/network';
 import { mark } from 'vs/workbench/contrib/notebook/common/notebookPerformance';
+import { IFileService } from 'vs/platform/files/common/files';
+import { AbstractResourceEditorInput } from 'vs/workbench/common/editor/resourceEditorInput';
 
 interface NotebookEditorInputOptions {
 	startDirty?: boolean;
 }
 
-export class NotebookEditorInput extends EditorInput {
+export class NotebookEditorInput extends AbstractResourceEditorInput {
 
 	static create(instantiationService: IInstantiationService, resource: URI, viewType: string, options: NotebookEditorInputOptions = {}) {
 		return instantiationService.createInstance(NotebookEditorInput, resource, viewType, options);
@@ -29,13 +31,11 @@ export class NotebookEditorInput extends EditorInput {
 
 	static readonly ID: string = 'workbench.input.notebook';
 
-	private readonly _name: string;
-
 	private _editorModelReference: IReference<IResolvedNotebookEditorModel> | null = null;
 	private _defaultDirtyState: boolean = false;
 
 	constructor(
-		public readonly resource: URI,
+		resource: URI,
 		public readonly viewType: string,
 		public readonly options: NotebookEditorInputOptions,
 		@INotebookService private readonly _notebookService: INotebookService,
@@ -43,10 +43,10 @@ export class NotebookEditorInput extends EditorInput {
 		@IFileDialogService private readonly _fileDialogService: IFileDialogService,
 		@IInstantiationService private readonly _instantiationService: IInstantiationService,
 		@ILabelService labelService: ILabelService,
+		@IFileService fileService: IFileService
 	) {
-		super();
+		super(resource, undefined, labelService, fileService);
 		this._defaultDirtyState = !!options.startDirty;
-		this._name = labelService.getUriBasenameLabel(resource);
 	}
 
 	override dispose() {
@@ -57,10 +57,6 @@ export class NotebookEditorInput extends EditorInput {
 
 	override get typeId(): string {
 		return NotebookEditorInput.ID;
-	}
-
-	override getName(): string {
-		return this._name;
 	}
 
 	override isDirty() {
@@ -80,6 +76,14 @@ export class NotebookEditorInput extends EditorInput {
 		}
 
 		return this._editorModelReference.object.isReadonly();
+	}
+
+	override isOrphaned() {
+		if (!this._editorModelReference) {
+			return super.isOrphaned();
+		}
+
+		return this._editorModelReference.object.isOrphaned();
 	}
 
 	override async save(group: GroupIdentifier, options?: ISaveOptions): Promise<IEditorInput | undefined> {
@@ -108,7 +112,7 @@ export class NotebookEditorInput extends EditorInput {
 			return undefined;
 		}
 
-		const dialogPath = this.isUntitled() ? await this._suggestName(this._name) : this._editorModelReference.object.resource;
+		const dialogPath = this.isUntitled() ? await this._suggestName(this.labelService.getUriBasenameLabel(this.resource)) : this._editorModelReference.object.resource;
 
 		const target = await this._fileDialogService.pickFileToSave(dialogPath, options?.availableFileSystems);
 		if (!target) {
@@ -175,6 +179,7 @@ export class NotebookEditorInput extends EditorInput {
 				return null;
 			}
 			this._register(this._editorModelReference.object.onDidChangeDirty(() => this._onDidChangeDirty.fire()));
+			this._register(this._editorModelReference.object.onDidChangeOrphaned(() => this._onDidChangeLabel.fire()));
 			if (this._editorModelReference.object.isDirty()) {
 				this._onDidChangeDirty.fire();
 			}
