+++ src/vs/workbench/contrib/output/browser/logViewer.ts
@@ -12,7 +12,7 @@ import { ITextResourceConfigurationService } from 'vs/editor/common/services/tex
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { AbstractTextResourceEditor } from 'vs/workbench/browser/parts/editor/textResourceEditor';
 import { IThemeService } from 'vs/platform/theme/common/themeService';
-import { ResourceEditorInput } from 'vs/workbench/common/editor/resourceEditorInput';
+import { TextResourceEditorInput } from 'vs/workbench/common/editor/textResourceEditorInput';
 import { URI } from 'vs/base/common/uri';
 import { ITextModelService } from 'vs/editor/common/services/resolverService';
 import { LOG_SCHEME } from 'vs/workbench/contrib/output/common/output';
@@ -22,9 +22,8 @@ import { IEditorService } from 'vs/workbench/services/editor/common/editorServic
 import { ITextFileService } from 'vs/workbench/services/textfile/common/textfiles';
 import { IFileService } from 'vs/platform/files/common/files';
 import { ILabelService } from 'vs/platform/label/common/label';
-import { IFilesConfigurationService } from 'vs/workbench/services/filesConfiguration/common/filesConfigurationService';
 
-export class LogViewerInput extends ResourceEditorInput {
+export class LogViewerInput extends TextResourceEditorInput {
 
 	static override readonly ID = 'workbench.editorinputs.output';
 
@@ -37,10 +36,8 @@ export class LogViewerInput extends ResourceEditorInput {
 		@ITextModelService textModelResolverService: ITextModelService,
 		@ITextFileService textFileService: ITextFileService,
 		@IEditorService editorService: IEditorService,
-		@IEditorGroupsService editorGroupService: IEditorGroupsService,
 		@IFileService fileService: IFileService,
-		@ILabelService labelService: ILabelService,
-		@IFilesConfigurationService filesConfigurationService: IFilesConfigurationService
+		@ILabelService labelService: ILabelService
 	) {
 		super(
 			URI.from({ scheme: LOG_SCHEME, path: outputChannelDescriptor.id }),
@@ -50,10 +47,8 @@ export class LogViewerInput extends ResourceEditorInput {
 			textModelResolverService,
 			textFileService,
 			editorService,
-			editorGroupService,
 			fileService,
-			labelService,
-			filesConfigurationService
+			labelService
 		);
 	}
 }
