+++ src/vs/workbench/contrib/performance/browser/perfviewEditor.ts
@@ -5,7 +5,7 @@
 
 import { localize } from 'vs/nls';
 import { URI } from 'vs/base/common/uri';
-import { ResourceEditorInput } from 'vs/workbench/common/editor/resourceEditorInput';
+import { TextResourceEditorInput } from 'vs/workbench/common/editor/textResourceEditorInput';
 import { ITextModelService, ITextModelContentProvider } from 'vs/editor/common/services/resolverService';
 import { ITextModel } from 'vs/editor/common/model';
 import { ILifecycleService, LifecyclePhase, StartupKindToString } from 'vs/workbench/services/lifecycle/common/lifecycle';
@@ -21,10 +21,8 @@ import { LoaderStats } from 'vs/base/common/amd';
 import { IProductService } from 'vs/platform/product/common/productService';
 import { ITextFileService } from 'vs/workbench/services/textfile/common/textfiles';
 import { IEditorService } from 'vs/workbench/services/editor/common/editorService';
-import { IEditorGroupsService } from 'vs/workbench/services/editor/common/editorGroupsService';
 import { ByteSize, IFileService } from 'vs/platform/files/common/files';
 import { ILabelService } from 'vs/platform/label/common/label';
-import { IFilesConfigurationService } from 'vs/workbench/services/filesConfiguration/common/filesConfigurationService';
 import { isWeb } from 'vs/base/common/platform';
 
 export class PerfviewContrib {
@@ -43,7 +41,7 @@ export class PerfviewContrib {
 	}
 }
 
-export class PerfviewInput extends ResourceEditorInput {
+export class PerfviewInput extends TextResourceEditorInput {
 
 	static readonly Id = 'PerfviewInput';
 	static readonly Uri = URI.from({ scheme: 'perf', path: 'Startup Performance' });
@@ -56,10 +54,8 @@ export class PerfviewInput extends ResourceEditorInput {
 		@ITextModelService textModelResolverService: ITextModelService,
 		@ITextFileService textFileService: ITextFileService,
 		@IEditorService editorService: IEditorService,
-		@IEditorGroupsService editorGroupService: IEditorGroupsService,
 		@IFileService fileService: IFileService,
-		@ILabelService labelService: ILabelService,
-		@IFilesConfigurationService filesConfigurationService: IFilesConfigurationService
+		@ILabelService labelService: ILabelService
 	) {
 		super(
 			PerfviewInput.Uri,
@@ -69,10 +65,8 @@ export class PerfviewInput extends ResourceEditorInput {
 			textModelResolverService,
 			textFileService,
 			editorService,
-			editorGroupService,
 			fileService,
-			labelService,
-			filesConfigurationService
+			labelService
 		);
 	}
 }
